@@ -1,3 +1,3 @@
 module Rollbar
-  VERSION = '3.5.2'.freeze
+  VERSION = '3.6.2'.freeze
 end
