@@ -288,7 +288,11 @@ describe Rollbar::Middleware::Sinatra, :reconfigure_notifier => true do
 
           expect(frames[-1][:locals]).to be_eql(locals[0])
 
-          expect(frames[-2][:method]).to be_eql('tap')
+          if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new('3.4.0')
+            expect(frames[-2][:method]).to be_eql('Kernel#tap')
+          else
+            expect(frames[-2][:method]).to be_eql('tap')
+          end
           if Gem::Version.new(RUBY_VERSION) >= Gem::Version.new('3.0.0')
             expect(frames[-2][:locals]).to be_nil
           else
